+++ src/AaveV3Optimism.sol
@@ -107,7 +107,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1
   address internal constant DAI_UNDERLYING = 0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1;
 
-  uint256 internal constant DAI_DECIMALS = 18;
+  uint8 internal constant DAI_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x82E64f49Ed5EC1bC6e43DAD4FC8Af9bb3A2312EE
   address internal constant DAI_A_TOKEN = 0x82E64f49Ed5EC1bC6e43DAD4FC8Af9bb3A2312EE;
@@ -127,7 +127,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x350a791Bfc2C21F9Ed5d10980Dad2e2638ffa7f6
   address internal constant LINK_UNDERLYING = 0x350a791Bfc2C21F9Ed5d10980Dad2e2638ffa7f6;
 
-  uint256 internal constant LINK_DECIMALS = 18;
+  uint8 internal constant LINK_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x191c10Aa4AF7C30e871E70C95dB0E4eb77237530
   address internal constant LINK_A_TOKEN = 0x191c10Aa4AF7C30e871E70C95dB0E4eb77237530;
@@ -148,7 +148,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x7F5c764cBc14f9669B88837ca1490cCa17c31607
   address internal constant USDC_UNDERLYING = 0x7F5c764cBc14f9669B88837ca1490cCa17c31607;
 
-  uint256 internal constant USDC_DECIMALS = 6;
+  uint8 internal constant USDC_DECIMALS = 6;
 
   // https://explorer.optimism.io/address/0x625E7708f30cA75bfd92586e17077590C60eb4cD
   address internal constant USDC_A_TOKEN = 0x625E7708f30cA75bfd92586e17077590C60eb4cD;
@@ -169,7 +169,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x68f180fcCe6836688e9084f035309E29Bf0A2095
   address internal constant WBTC_UNDERLYING = 0x68f180fcCe6836688e9084f035309E29Bf0A2095;
 
-  uint256 internal constant WBTC_DECIMALS = 8;
+  uint8 internal constant WBTC_DECIMALS = 8;
 
   // https://explorer.optimism.io/address/0x078f358208685046a11C85e8ad32895DED33A249
   address internal constant WBTC_A_TOKEN = 0x078f358208685046a11C85e8ad32895DED33A249;
@@ -190,7 +190,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x4200000000000000000000000000000000000006
   address internal constant WETH_UNDERLYING = 0x4200000000000000000000000000000000000006;
 
-  uint256 internal constant WETH_DECIMALS = 18;
+  uint8 internal constant WETH_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0xe50fA9b3c56FfB159cB0FCA61F5c9D750e8128c8
   address internal constant WETH_A_TOKEN = 0xe50fA9b3c56FfB159cB0FCA61F5c9D750e8128c8;
@@ -211,7 +211,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x94b008aA00579c1307B0EF2c499aD98a8ce58e58
   address internal constant USDT_UNDERLYING = 0x94b008aA00579c1307B0EF2c499aD98a8ce58e58;
 
-  uint256 internal constant USDT_DECIMALS = 6;
+  uint8 internal constant USDT_DECIMALS = 6;
 
   // https://explorer.optimism.io/address/0x6ab707Aca953eDAeFBc4fD23bA73294241490620
   address internal constant USDT_A_TOKEN = 0x6ab707Aca953eDAeFBc4fD23bA73294241490620;
@@ -232,7 +232,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x76FB31fb4af56892A25e32cFC43De717950c9278
   address internal constant AAVE_UNDERLYING = 0x76FB31fb4af56892A25e32cFC43De717950c9278;
 
-  uint256 internal constant AAVE_DECIMALS = 18;
+  uint8 internal constant AAVE_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0xf329e36C7bF6E5E86ce2150875a84Ce77f477375
   address internal constant AAVE_A_TOKEN = 0xf329e36C7bF6E5E86ce2150875a84Ce77f477375;
@@ -253,7 +253,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x8c6f28f2F1A3C87F0f938b96d27520d9751ec8d9
   address internal constant sUSD_UNDERLYING = 0x8c6f28f2F1A3C87F0f938b96d27520d9751ec8d9;
 
-  uint256 internal constant sUSD_DECIMALS = 18;
+  uint8 internal constant sUSD_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x6d80113e533a2C0fe82EaBD35f1875DcEA89Ea97
   address internal constant sUSD_A_TOKEN = 0x6d80113e533a2C0fe82EaBD35f1875DcEA89Ea97;
@@ -274,7 +274,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x4200000000000000000000000000000000000042
   address internal constant OP_UNDERLYING = 0x4200000000000000000000000000000000000042;
 
-  uint256 internal constant OP_DECIMALS = 18;
+  uint8 internal constant OP_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x513c7E3a9c69cA3e22550eF58AC1C0088e918FFf
   address internal constant OP_A_TOKEN = 0x513c7E3a9c69cA3e22550eF58AC1C0088e918FFf;
@@ -294,7 +294,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x1F32b1c2345538c0c6f582fCB022739c4A194Ebb
   address internal constant wstETH_UNDERLYING = 0x1F32b1c2345538c0c6f582fCB022739c4A194Ebb;
 
-  uint256 internal constant wstETH_DECIMALS = 18;
+  uint8 internal constant wstETH_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0xc45A479877e1e9Dfe9FcD4056c699575a1045dAA
   address internal constant wstETH_A_TOKEN = 0xc45A479877e1e9Dfe9FcD4056c699575a1045dAA;
@@ -315,7 +315,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0xc40F949F8a4e094D1b49a23ea9241D289B7b2819
   address internal constant LUSD_UNDERLYING = 0xc40F949F8a4e094D1b49a23ea9241D289B7b2819;
 
-  uint256 internal constant LUSD_DECIMALS = 18;
+  uint8 internal constant LUSD_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x8Eb270e296023E9D92081fdF967dDd7878724424
   address internal constant LUSD_A_TOKEN = 0x8Eb270e296023E9D92081fdF967dDd7878724424;
@@ -336,7 +336,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0xdFA46478F9e5EA86d57387849598dbFB2e964b02
   address internal constant MAI_UNDERLYING = 0xdFA46478F9e5EA86d57387849598dbFB2e964b02;
 
-  uint256 internal constant MAI_DECIMALS = 18;
+  uint8 internal constant MAI_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x8ffDf2DE812095b1D19CB146E4c004587C0A0692
   address internal constant MAI_A_TOKEN = 0x8ffDf2DE812095b1D19CB146E4c004587C0A0692;
@@ -356,7 +356,7 @@ library AaveV3OptimismAssets {
   // https://explorer.optimism.io/address/0x9Bcef72be871e61ED4fBbc7630889beE758eb81D
   address internal constant rETH_UNDERLYING = 0x9Bcef72be871e61ED4fBbc7630889beE758eb81D;
 
-  uint256 internal constant rETH_DECIMALS = 18;
+  uint8 internal constant rETH_DECIMALS = 18;
 
   // https://explorer.optimism.io/address/0x724dc807b04555b71ed48a6896b6F41593b8C637
   address internal constant rETH_A_TOKEN = 0x724dc807b04555b71ed48a6896b6F41593b8C637;
